add change in project

+++ node/controller/productController.js
@@ -4,21 +4,24 @@ const getAllProduct=async(req,res)=>{
 
     const Products=await Product.find()
     if(!Products?.length){
-        return res.status(400).json({massage:'no Product'})
+        return res.status(400).json({massage:'no Product found'})
     }
+
     res.json(Products)
+
 }
 const creataNewProduct=async(req,res)=>{
+
     const{name,description,price,quantity}=req.body
-    
     if(!name||!price||!req.files){
         return res.status(400).json({message:'filds are required'})
     }
 
-    let imageURL=[]
+    let imageURL=[] 
     req.files.forEach(element => {
         imageURL.push(element.path)
     });
+
     const product=await Product.create({name,imageURL,description,price,quantity})
     if(product){
         return res.status(201).json(product)
@@ -26,20 +29,25 @@ const creataNewProduct=async(req,res)=>{
     else{
         return res.status(400).json({message:'invalid Product'}) 
     }
+
 }
 const updateProduct=async(req,res)=>{
+   
     const{_id,name,description,price,quantity}=req.body 
     if(!_id||!name||!price){
         return res.status(400).json({message:'fields are required'})
     }
+    
     const product=await Product.findById(_id).exec()
     if(!product){
         return res.status(400).json({message:'Product not found'})
     }
+    
     const imageURL=[]
     req.files?.forEach(element => {
         imageURL.push(element.path)
     });
+
     product.name=name
     product.imageURL=imageURL.length>0?imageURL:product.imageURL
     product.price=price
@@ -47,23 +55,36 @@ const updateProduct=async(req,res)=>{
     product.quantity=quantity
     const updateProduct=await product.save()
 
-    res.json(updateProduct)
+    res.json(updateProduct) 
 }
+
 const deletProduct=async(req,res)=>{
+    
     const{_id}=req.body
+    if(!_id){
+        return res.status(400).json({message:'field are required'})
+    }
+
     const product=await Product.findById(_id).exec()
     if(!product){
         return res.status(201).json({message:'Product not found'})
     }
+    
     const result=await product.deleteOne()
     res.json(result)
 }
 const getProductbyId = async (req, res) => {
+    
     const { id } = req.params
+    if(!id){
+        return res.status(400).json({message:'field are required'})
+    }
+
     const product = await Product.findById(id).lean()
     if (!product) {
         return res.status(400).json({ message: 'no product found' })
     }
+
     res.json(product)
 
 }
